MP3 files playback works
- player.PlayerImpl class added as an implementation of player.Player interface
- player.PlayerThread class added as a main controlling thread for all operations related to the player

+++ src/main/java/org/luwrain/app/player/Listener.java
@@ -43,7 +43,7 @@ class Listener  implements org.luwrain.player.Listener
 	luwrain.runInMainThread(()->area.onNewTrack(trackNum));
     }
 
-    @Override public void onTrackTime(int sec)
+    @Override public void onTrackTime(long sec)
     {
 	luwrain.runInMainThread(()->area.onTrackTime(sec));
     }
